packnet-pytorch/requirements.txt
torch>=1.10.0
numpy
scipy
pillow
tqdm
matplotlib 